calender_app/requirements.txt
langchain
langchain-core
langchain-google-genai
python-dotenv
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib